Rename lotr_lcg/types.py to card_entry.py to fix stdlib shadowing crash
python plugins/lotr_lcg/fetch.py <anything> crashed immediately on every
invocation: Python auto-prepends a script's own directory to sys.path,
so plugins/lotr_lcg/types.py shadowed the real stdlib types module for
the whole process. click's own import chain (enum -> functools -> types)
hit the plugin's file instead and failed with ImportError before fetch.py
ever ran.

This wasn't caught by the existing test suite because it imports the
module as a package (plugins.lotr_lcg.types), by which point pytest and
other machinery have already imported and cached the real stdlib types,
masking the collision. It only reproduces when fetch.py is run directly
as a script, which is the only way the plugin's own README documents.
