+++ plugins/lotr_lcg/card_entry.py
